[REF] Refactor Anomaly Detection Module into Submodules by Algorithm Family (#2694)
* Refactor Anomaly Detection Module into Submodules by Algorithm Family

* updated documentation and references

* implemented suggested changes

* minor changes

* added headers for remaining algorithm family

* removing tree-based header
