+++ aeon/anomaly_detection/__init__.py
@@ -1,31 +1,7 @@
 """Time Series Anomaly Detection."""
 
 __all__ = [
-    "CBLOF",
-    "COPOD",
-    "DWT_MLEAD",
-    "IsolationForest",
-    "KMeansAD",
-    "LeftSTAMPi",
-    "LOF",
-    "MERLIN",
-    "OneClassSVM",
-    "ROCKAD",
-    "PyODAdapter",
-    "STOMP",
-    "STRAY",
+    "BaseAnomalyDetector",
 ]
 
-from aeon.anomaly_detection._cblof import CBLOF
-from aeon.anomaly_detection._copod import COPOD
-from aeon.anomaly_detection._dwt_mlead import DWT_MLEAD
-from aeon.anomaly_detection._iforest import IsolationForest
-from aeon.anomaly_detection._kmeans import KMeansAD
-from aeon.anomaly_detection._left_stampi import LeftSTAMPi
-from aeon.anomaly_detection._lof import LOF
-from aeon.anomaly_detection._merlin import MERLIN
-from aeon.anomaly_detection._one_class_svm import OneClassSVM
-from aeon.anomaly_detection._pyodadapter import PyODAdapter
-from aeon.anomaly_detection._rockad import ROCKAD
-from aeon.anomaly_detection._stomp import STOMP
-from aeon.anomaly_detection._stray import STRAY
+from aeon.anomaly_detection.base import BaseAnomalyDetector
