+++ aeon/anomaly_detection/distance_based/_left_stampi.py
@@ -44,7 +44,7 @@ class LeftSTAMPi(BaseAnomalyDetector):
     Internally,this is applying the incremental approach outlined below.
 
     >>> import numpy as np # doctest: +SKIP
-    >>> from aeon.anomaly_detection import LeftSTAMPi  # doctest: +SKIP
+    >>> from aeon.anomaly_detection.distance_based import LeftSTAMPi  # doctest: +SKIP
     >>> X = np.random.default_rng(42).random((10))  # doctest: +SKIP
     >>> detector = LeftSTAMPi(window_size=3, n_init_train=3)  # doctest: +SKIP
     >>> detector.fit_predict(X)  # doctest: +SKIP
