+++ aeon/anomaly_detection/distance_based/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Distance based test code."""
