+++ aeon/anomaly_detection/distribution_based/__init__.py
@@ -1,0 +1,9 @@
+"""Distribution based Time Series Anomaly Detection."""
+
+__all__ = [
+    "COPOD",
+    "DWT_MLEAD",
+]
+
+from aeon.anomaly_detection.distribution_based._copod import COPOD
+from aeon.anomaly_detection.distribution_based._dwt_mlead import DWT_MLEAD
