+++ aeon/anomaly_detection/distribution_based/_dwt_mlead.py
@@ -78,7 +78,7 @@ class DWT_MLEAD(BaseAnomalyDetector):
     Examples
     --------
     >>> import numpy as np
-    >>> from aeon.anomaly_detection import DWT_MLEAD
+    >>> from aeon.anomaly_detection.distribution_based import DWT_MLEAD
     >>> X = np.array([1, 2, 3, 4, 1, 2, 3, 3, 2, 8, 9, 8, 1, 2, 3, 4], dtype=np.float64)
     >>> detector = DWT_MLEAD(
     ...    start_level=1, quantile_boundary_type='percentile', quantile_epsilon=0.01
