+++ aeon/anomaly_detection/distribution_based/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Distribution based test code."""
