+++ aeon/anomaly_detection/outlier_detection/__init__.py
@@ -1,0 +1,11 @@
+"""Time Series Outlier Detection."""
+
+__all__ = [
+    "IsolationForest",
+    "PyODAdapter",
+    "STRAY",
+]
+
+from aeon.anomaly_detection.outlier_detection._iforest import IsolationForest
+from aeon.anomaly_detection.outlier_detection._pyodadapter import PyODAdapter
+from aeon.anomaly_detection.outlier_detection._stray import STRAY
