+++ aeon/anomaly_detection/outlier_detection/_pyodadapter.py
@@ -59,7 +59,9 @@ class PyODAdapter(BaseAnomalyDetector):
     --------
     >>> import numpy as np
     >>> from pyod.models.lof import LOF  # doctest: +SKIP
-    >>> from aeon.anomaly_detection import PyODAdapter  # doctest: +SKIP
+    >>> from aeon.anomaly_detection.outlier_detection import (
+    ...     PyODAdapter
+    ... ) # doctest: +SKIP
     >>> X = np.random.default_rng(42).random((10, 2), dtype=np.float64)
     >>> detector = PyODAdapter(LOF(), window_size=2)  # doctest: +SKIP
     >>> detector.fit_predict(X, axis=0)  # doctest: +SKIP
