+++ aeon/anomaly_detection/outlier_detection/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Outlier based test code."""
