+++ aeon/anomaly_detection/whole_series/__init__.py
@@ -1,0 +1,7 @@
+"""Whole Time Series Anomaly Detection."""
+
+__all__ = [
+    "ROCKAD",
+]
+
+from aeon.anomaly_detection.whole_series._rockad import ROCKAD
