+++ aeon/anomaly_detection/whole_series/_rockad.py
