+++ aeon/anomaly_detection/whole_series/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Whole series anomaly detection tests."""
